backend/requirements.txt
fastapi
uvicorn
google-auth-oauthlib
google-api-python-client
python-multipart
requests
fpdf
Pillow
python-dotenv
google-generativeai
